requirements.txt
fastapi
uvicorn
requests
pydantic
sentence-transformers
faiss-cpu
numpy
beautifulsoup4
playwright
markdownify
python-multipart

